I have changed a lot of things this time and I didn't commit before trying things out. It looks like I am at a structure that is very similar to the final version now.
First of all, the whole structure is different from the last commit. I
changed base structs, including the names from element -> unit from
container -> combo etc..

I also created two UiUnits. UiEmptyItem, which doesn't do anything. And
UiSplitCombo wich makes it able to split a windo in multiple children in
a direction, with either fixed size for the element, or relative to the
space that is left. UiSplitCombo has some bugs (as division by 0) and is
very unclean but I will make sure to improve on them the next commits.

I also added another example. That is a responsive window wich is
supposed to demonstrate how a sidebar should look like using this
librar using this libraryy.

I tried to add a Readme Markdown file too.